move block operations to own module
